PB-1259: allow bgLayer modification through URL
Issue: when the application was already loaded, it was no longer possible to set the background layer to the void layer.

Cause: the url parser validation checked for the 'void' string which is not in the background layers, and the setter only accepted null as a valid option.

Fix : we allowed the 'void' string value in both cases