PB-1260: handle legacy value 'voidLayer'
Issue: In the old mapviewer, it was possible to have the bgLayer parameter to equals 'voidLayer', which is equivalent to the current 'void' parameter. However, it caused issues where it displayed errors and refused to show the correct "layer".

Cause: we did not handle the parameter at all in the legacy parser.

Fix: when bgLayer is set to 'voidLayer' while navigating a legacy route, we handle it correctly and transform it to 'void'