PB-1176: Prepare build output for CD Pipeline
In order to simplify the CD Pipeline put the different build output based on the
same staging names that we have aka; dev, int and prod.

Also added more information to the info.json file that will be used as Pipeline
trigger.

Changed the info.json generator to make full use of nodejs async functionality.