PB-1254: fix embed test
Issue : the tests for the embed links checked strings as they are without any further processing. This meant that if the order of the parameters changed, it would make the test fail.

Fix : We update the tests with a deep comparison between two 'URLSearchParams' made from the variable parameter and the validation string