PB-1246 : add location search entry title to search bar when selected
users were confused that the full address of what they selected in the search result wasn't properly shown in the UI.

This only adds the title of such entries to the search bar, without triggering a new search. This means the search results aren't cleared/emptied when a location is selected.