PB-1176: Fix cesium static assets location
During the build the cesium static assets were put into the root of the application.
This was an issue for our deployment strategy were we have all assets in a
versioned folder and were we first deploy the versioned folder before deploying
the entry point index.html that point to the versioned folder. This allow for
a pseudo atomic deployment.

So now all the cesium static assets have been placed into VERSION/cesium/ path.