PB-1296 : share composable between GPX, KML and GeoJSON in Cesium
gather all the logic to manage a primitive based data source into a composable and use it in the 3 cases.

Adding a clampToGround option for KML layers, making it possible to have it "on top" of the terrain (it will ignore any Z value if set so).

Tweaking a bit the style for Swissnames to have a buffer/offset between the ground and the text (with an anchor line/post, pointing to the ground)