PB-1327 :  i18n returns non-translated input if not found in i18n files
instead of failing and raising an error while using t(...), vue-i18n now gives back the "key" or text it received unchanged. This way we can call translation tool on input from other component without worrying on raising errors.

The DropdownButton now translates anything it receives, making it possible to give it either keys, or translated text directly.

Removing the class DropdownItem and replacing it with a typedef instead. We now know that passing classes with get/set to Vuex is a bad practice, and we should avoid it.