PB-1151: Notify user when imported file is not all in extent
Issue : When a user import a file which is partially out of bounds, we want to notify the user that the file is not completely within bounds, and that some functionalities might not be available.

Fix : we add a flag in the store, and the state of the flag will tell us if we need to show the warning message or the valid message.