PB-1384 : modularize proj, number and coordinate utils
in prep work to the elevation profile module, I started modularizing what will be used by both the viewer and the future elevation profile component.

I used the opportunity of having a smaller scoped project to switch to Typescript entirely all the utilities.
Migration to ESLint9 was also done here, meaning there are some changes because of new linter rules. Some ESLint8 plugins weren't available in ESLint9 so I had to find equivalents, but they don't have exactly the same way of linting

as we are running a preview on a "from scratch" CI instance, geoadmin package has never been built (and added to the local npm registry for that matter).
Building the lib with the same target in mind before running the preview script on the viewer