PB-1384 Fix eslint/prettier relationship
Provide a combination of eslint and prettier that sort of respects
vue/max-attributes-per-line.

Configure eslint and prettier so that the attributes in HTML code are
put on single lines if there are more than two. The formatting is done
entirely by Prettier, since [eslint doesn't want to be a formatting tool
anymore](https://eslint.org/blog/2023/10/deprecating-formatting-rules/)

This also disables the running of prettier from eslint, which is
the [recommended
way](https://github.com/vuejs/eslint-config-prettier?tab=readme-ov-file#use-separate-commands-for-linting-and-formatting)