PB-1436 Switch to using pnpm
Adding a workspace configuration with a catalog to be used from all our
packages. This will require to run `pnpm install` on the top level and
it will install all required packages in our modularized packages.

The catalog in the workspace configuration helps having the same version
across all our sub-packages.