Don't render invisible layer catalogue items
Don't render the invisible layer catalogue items. Only render them when
the nested navigation is un-collapsed.
This produces quite a performance gain, as many sub-sequent components
don't need rendering. This includes <TextTruncate>, where each of them
creates a ResizeObserver on the DOM. Without this change we have ~1100
resize observers, with the change there are only ~250.

I will investigate those ~250 maybe in another attempt, some of them
might be resolved when switching to the new tooltip.