PB-1137 : create a full HTML confederation text (for the logo)
We will want to re-size that part of the UI when we print, and using a flatten PNG for that part of the logo will be a problem.
The Swiss flag is already a SVG, so I thought why not change the text to also be rendered by HTML/CSS.

Importing both Frutiger Light and Bold to the project, and switching the whole code base to Frutiger Light.