PB-1004 : only use right click to trigger "delete point" action
users on touch devices could trigger a delete by double tapping on the last point they were drawing, which wasn't ideal.

As I've run out of ideas on how to fix this on mobile, I've simply removed the possibility to delete a point on mobile as we lack "gestures" to achieve it.