REWORD Introducing the first layers in the new module
Replace GeoAdminWMTSLayer class, fix everything that relied on it

Do WMS Layer work

Fix Layer Features for WMS and WMTS

Fix some stuff from the previous commits

Fix imporToolMaps tests

The menu wasn't working properly anymore in the tests. Could be due to
the presence of the vue devtools.

Try to typescriptify the existing layers - maybe try to get rid of some
of this

Attempt to do something with utils

implementing geojson

squash this