PB-1432 Move data-cy attribution away from props
Due to the nature of GeoadminTooltip an element was wrapped with that
new component. This lead to the situation where cypress couldn't select
the correct element.
Now instead of working around this with a prop to the component lets
change the selector in cypress and add the attribute programmatically
just before creating the ol/Overlay