PB-1137 : adding print view
so that we may start developing our new headless print backend

Resolution from @geoadmin/coordinates are now expressed in steps, where each step can have a corresponding zoom level or not.
This is needed with LV95, where we have some gaps in zoom level, while there is a resolution (between zoom 9 and 10 there's an extra resolution step)

Also making our coordinate system calculate resolution without threshold (especially for LV95) so that it can then be used to calculate a mercator zoom level. Will be very useful when we will be switching to using VectorTiles (and Mercator on-the-fly reprojection)