PB-1493 : Add a way to stop the zoom when in embed mode
Issue : We wanted to add a way for people to have an iframe which did not zoom
automatically when we were scrolling the page on which they were embedded. We
also did not want to alter the default behavior, as we did not want thousands of mails

Fix: We added a new URL parameter, and put it in the store, to alter the behavior
of the scroll on embed views when it is active. We also added a button to toggle it when
creating a new iframe.

Added test for scrolling embedded preview & make modal window size dynamic