PB-878: removing store usage in print.api.js
Issue: the store is not meant to be used outside of vue components or composables.
Using it within the `print.api.js` file caused issues when testing.

Fix : we give the print extent as a parameter throughout the print.api.js file,
and get it from the store in the usePrint compoasable file.