PB-1605 : detect COG projection based on name when no EPSG found
Some COGs we provide aren't encoded with an EPSG number in the corresponding field, but with a specific number telling us it is a "user defined CRS".
In order to be able to detect such COGs, I've added a technical name to both LV95 and LV03, which will be used to detect COGs that are in such format.