PB-1522 : DPI responsive print margin
We will move the generation of margin in the viewer, instead of dealing with that on the backend. This is because it's way easier to keep the "true-to-scale" when shrinking the map viewport (and let it recalculate everything) instead of resizing the flatten map and thus changing the scale of the image.

Using CSS padding instead of margin, so that we respect the A4 size we previously calculated (if we use margin, it will extend the "page" size and break the A4 layout)