PB-1674: Fix tooltip creating too much height with map buttons
The geolocation button has a "linking background" with the compass button (which isn't visible when pointing north).

Since modularization, this background is swelling under the button

also fixing an issue with the "share link" input, that was not expending as intended since the tooltip modularization.

I also took the opportunity to only color the border of the input when a warning is shown, and moved the tooltip with the warning text to the copy button