PB-1670: clarify lack of precision of mouse tracked LV03 coordinates
As pointed out by some of our users, the dynamically generated coordinates on the bottom right part of the map with the mouse are not precise when dealing with LV03.
They are generated without taking into consideration the real deformation between LV95 and LV03, and are calculated based on a simple matrix reprojection to be more dynamic (we are not using the REFRAME service after each mouse move, that would be overkill).