PB-1579: add support for local BigTIFF
we were missing the little endian signature for BigTIFF (which is different than normal TIFFs) to be able to parse a local BigTIFF file.

It was working fine with URLs already because we do not test against little endian signature in this case, and let GeoTIFF.js library do its magic, so nothing to change there.