PB-1697: limiting floating window's height
Some users that use the browser zoom to enlarge the UI have issues with the SimpleWindow component (especially when it's used by the feedback component) because it grows too heigh, and overflow the header (some even saying the X button of the window was out of screen).

So to fix that, I've limited the height of all simple windows to 2/3 of the height of the screen. This was tested with a browser zoom of 200% and having the "report a problem" window opened.