PB-1754: adding a lib that can conditionally build part of the JS
as Service Worker is disabled when testing with Cypress, the component that was importing part of the SW API function was failing to build when running Cypress. With this conditional build, it can now be emptied (and especially not import the SW part) when we are running Cypress.

This lib needed rollup v4+ so I've set it as such in the overrides of the package.json of the mapviewer package

Also adding TypeScript types for all env variables declared through Vite and .env files