PB-1814: Fix download print issue since Service-Worker introduction
Removes the `file-saver` dependency and implements a custom `downloadFile` function to handle file downloads.

This change ensures compatibility and addresses potential issues related to browser-specific behavior while Service-Worker is caching our app.
Prior to this change, opening the download URL the way it was was leading to a white page where it looked like Service-Worker was trying to inject the app code there (and it failed), no PDF was downloaded