PB-1895: Stop opening drawing mode at startup when the KML is not valid
Issue: When starting the mapviewer with an edit link for a KML that was deleted,
the drawing module is open and can't save to a new KML as the presence of the initial
admin ID is interfering with that function.

Fix: We now ignore the admin ID and display an error when someone is trying to
load a KML that was deleted.

Also: We altered the structure of messages a bit so that the same error given by
the same source is not shown, but the same error from a different source would be displayed

Also Also : We added some tests to check that the message equality function was working
correctly.

Also Also Also : We added an acknowledgement system which stops the same source from
sending the same errors multiple times