PB-1383: refactor 3d e2e tests to typescript
Issue: We want to change the whole codebase to typescript, tests included

Fix: We change the tests to typescripts. We remove some variables that were using non-existent attributes, and we states that we're expecting errors from files that are not yet set in typescript. Those "expect errors" lint exception will yell at us when the typescript conversion has been correctly made :)

we also update the tsconfig to explicitly accept some functionalities