PB-1383: remove Cesium getViewer util function
as it is not authorized to use an `inject` (from Vue) outside of a setup block, this wasn't compiling/transpiling correctly.

I've instead opted to use the same approach as with the OL components, meaning we receive a reference to the "map" directly, instead of a getter function.
Adapting the code everywhere the util was used to manage an undefined viewer (and fail somewhat gracefully in the case it is undefined)