PB-1383: re-arrange geolocation store into multiple files
The idea was to have one action per file, and share the reference to the store anyway (through this).
I found this solution on Pinia GitHub tracker https://github.com/vuejs/pinia/discussions/983#discussioncomment-2063235

a bit verbose/boilerplate, but that makes it way clearer what is happening in an action, especially now that we want to move all the "complex interactions" the plugins were doing into the store itself.