PB-2037: fix print scales not loading
Issue : When loading the print menu for the first time, the scales and the background preview of the layout would not show. Toggling the print menu would then show the scales and the background preview instead of closing the menu, and the menu would then work as intended.

Cause : When toggling the menu for the first time, we wanted to set a watcher within the toggle menu function which would catch when the print layouts changed to load the layouts and select one. This managed to not fire as intended, and simultaneously shortcut the already existing watcher, resulting in a state where the menu was open, but not considered to be open, which meant the scales would not be computed.

Fix: We keep loading the layouts if they are not loaded, but we no longer wait on a watcher that's not firing up to show the menu.