PB-2036: split previous VueX plugin into two
one that deals with route changes, and one that deals with store changes.

this is now required as Pinia will initialize one plugin per store instance (meaning one for the app store, one for the cesium store, one for the layers store, etc...) and we can't be listening in all these instances of the plugin to the router changes and react to it (that would make 13+ reaction)

So the logic that was dealing with URL changes has been moved to a router plugin.