PB-1383: infinite redirect during tests with fake Kataster layer
Issue: When parsing external layers, if we encountered a layer with a sublayer, and both had an identical name, we would enter an infinite loop as we would be finding the base layer over and over again when going through the child to get valid layers.

Temporary fix: If we have sublayers, we only go through the loop if those sublayers have sublayers.

Next step : A more elegant solution would be preferable, as this can still fail if the sublayer has the same name as the base layer and has sub layers of its own.