PB-2064: remove most, if not all, enums from the code
I've some suspicion that Cypress' TS handling is struggling with our enum, so I wanted to test out one CI run without them.

I've also re-ordered a bit some types from the API folder, and from the store.
Icons and Features type are now in a dedicated file (I wanted to test if the import of the whole file was the problem with Cypress, and it looks better organized this way anyway, so why not keep it so)
Store type files are also renamed index.ts, so the import of types looks a bit more logical (and shorter)

I've locked pretter-plugin-jsdoc to the version 1.5.0 as the newer version creates some infinite loop issues