PB-1383: adding an ESLint plugin/rule to sort TS imports
splitting type imports from normal imports

Prettier is also now "monorepo"-wide, as it was easier to set it up this way because of the way plugins are handled by Prettier. We cannot specify them any other way than with a string identifier, that will then be used to create an import statement. The issue in a monorepo is that, when this import statement is "called", the node_modules aren't the one were we created the config, forcing us to add ALL plugins and prettier to ALL package that use it. Having it set up once and for all at the root of the monorepo is a good enough solution (there was not package specific configuration)